refactor(scaffold): real template files; channel bundles live with the channel
The files init/add scaffold lived as escaped TS string literals (scaffold-templates.ts), so
nested code escaped its backticks/backslashes (\\s+, the send-tool's string concatenation,
every nested template). Make them real files; a channel's bundle lives WITH the channel, so a
channel owns its starter kit and could later ship as its own package.

- Base workspace kit → src/scaffold/templates/ (AGENTS.md, env.example, gitignore,
  fastagent.config.mjs, skills/house-style/SKILL.md, tools/word-count.ts).
- Each channel's bundle → src/channels/<kind>/scaffold/ (channel.ts + companion tools, e.g.
  telegram's telegram-send.ts). `add <kind>` reads the bundle by convention: channel.ts →
  channels/<kind>.ts, any other .ts → tools/<name>. env/steps stay a small typed registry in
  add-channel.ts (a manifest in the bundle is the future-extraction step; YAGNI now).
- Drop the named-const layer; templates.ts is now just readers (baseTemplate / channelTemplate
  / channelBundleFiles) plus the parametric packageJson / toPackageName.
- AGENTS.md is one real minimal persona (no word-count nudge — the tool self-describes; one
  static file for both --minimal and complete). `add` does NOT touch AGENTS.md: it only drops
  new files and appends to .env.example, never mutating the author's identity file.
- templates/ + channels/*/scaffold/ are excluded from tsc + biome (they import the published
  package, not this source) and copied into dist by scripts/copy-scaffold-assets.mjs.
- The de-escaped send-tool reads as clean template literals; word-count's regex is /\s+/.

Verified end-to-end from dist: init writes the minimal persona + clean word-count; add telegram
drops channels/telegram.ts + tools/telegram-send.ts and leaves AGENTS.md untouched; npm pack
ships all templates under dist. typecheck/lint/248 tests green.
