+++ core/src/scaffold/templates/AGENTS.md
@@ -1,0 +1,5 @@
+# Assistant
+
+You are a concise, helpful assistant. Answer directly and skip filler.
+
+When the user asks you to write or edit prose, consult the house-style skill first.
