+++ core/src/scaffold/templates/env.example
@@ -1,0 +1,19 @@
+# Environment for this agent. Copy to .env (gitignored) and uncomment what you need.
+# Everything here is OPTIONAL — the defaults work without a .env.
+
+# --- Model auth ---
+# The default model (openai-codex) signs in with OAuth, not an API key: run `fastagent login` once.
+# Switch to an API-key provider? Set its key here — the variable name is provider-specific
+# (e.g. OPENAI_API_KEY, ANTHROPIC_API_KEY). Run `fastagent models` to see available specs.
+# OPENAI_API_KEY=
+# ANTHROPIC_API_KEY=
+
+# --- Model selection (overrides fastagent.config) ---
+# Precedence: --model flag > FASTAGENT_MODEL > config.
+# FASTAGENT_MODEL=openai-codex/gpt-5.5
+
+# --- Serving (fastagent start) ---
+# Port precedence: --port > PORT > config.http.port > 8787
+# PORT=8787
+# Where conversations persist (default: ./fastagent-sessions)
+# FASTAGENT_SESSIONS_DIR=./fastagent-sessions
