+++ core/src/scaffold/templates/gitignore
@@ -1,0 +1,8 @@
+# secrets — never commit (kept out of git and any deploy copy)
+.env
+
+# dependencies (reinstalled at deploy)
+node_modules/
+
+# fastagent machine state (dev/start sessions)
+.fastagent/
