+++ core/src/scaffold/templates/skills/house-style/SKILL.md
@@ -1,0 +1,9 @@
+---
+name: house-style
+description: The house writing style. Consult before writing or editing any prose for the user.
+---
+# House style
+
+- Prefer short sentences and the active voice.
+- Avoid marketing adjectives ("seamless", "powerful", "robust").
+- Lead with the answer; put caveats after.
