+++ core/tsconfig.json
@@ -13,5 +13,9 @@
     "verbatimModuleSyntax": true,
     "allowImportingTsExtensions": true
   },
-  "include": ["src", "test"]
+  "include": ["src", "test"],
+  // Scaffold payload: real template files copied verbatim into a new workspace (base kit under
+  // src/scaffold/templates, per-channel bundles under src/channels/*/scaffold). They import the
+  // published @kid7st/fastagent, not this source, so they are data here, not part of the build.
+  "exclude": ["src/scaffold/templates", "src/channels/*/scaffold"]
 }
